fix: address code review issues in foundation library
- value.h: fix memcpy size sizeof(int64_t) -> sizeof(double) in Value(double)
- timecodefunctions.cpp: fix secs = total_seconds % 60 (was wrong for hours > 0)
- timecodefunctions.cpp: strip sign from first token to avoid double-negative
- timerange.cpp: fail-fast on null timebase (clear list instead of just warning)
- timerange.cpp: guard chunk_size <= 0 in Split() to prevent div-by-zero
- stringutils.cpp: add missing <cstdio> include for vsnprintf
- stringutils.h: add missing <cctype> include for std::isspace
- samplebuffer.cpp: guard speed==0/NaN/inf in SampleBuffer::speed()
- tests.cpp: fix %lu -> %zu for size_t (MSVC x64 portability)
- color.cpp: use if/else if chain in toHsl() to avoid hue overwrite

+++ foundation/include/arcvideo/foundation/util/value.h
@@ -64,7 +64,7 @@ public:
   Value(double v)
   {
     data_.resize(sizeof(double));
-    memcpy(data_.data(), &v, sizeof(int64_t));
+    memcpy(data_.data(), &v, sizeof(double));
     type_ = FLOAT;
   }
